chore: move duplicated TestResource to test/test-resource.ts
