+++ test/aws/notify/event-bus.test.ts
@@ -9,14 +9,7 @@ import {
   kmsKey,
   iamRolePolicy,
 } from "@cdktf/provider-aws";
-import {
-  Lazy,
-  TerraformMetaArguments,
-  TerraformOutput,
-  TerraformResource,
-  Testing,
-} from "cdktf";
-import { Construct } from "constructs";
+import { Lazy, TerraformOutput, Testing } from "cdktf";
 import { AwsStack, Arn } from "../../../src/aws";
 import * as kms from "../../../src/aws/encryption";
 import * as iam from "../../../src/aws/iam";
@@ -24,6 +17,7 @@ import * as sqs from "../../../src/aws/notify";
 import "cdktf/lib/testing/adapters/jest";
 import { EventBus } from "../../../src/aws/notify/event-bus";
 import { Template } from "../../assertions";
+import { TestResource } from "../../test-resource";
 
 const environmentName = "Test";
 const gridUUID = "123e4567-e89b-12d3";
@@ -778,45 +772,3 @@ describe("event bus", () => {
     );
   });
 });
-
-export enum TestProviderMetadata {
-  TYPE = "test",
-}
-export interface TestResourceConfig extends TerraformMetaArguments {
-  readonly properties?: { [name: string]: any };
-}
-export class TestResource extends TerraformResource {
-  public static readonly tfResourceType: string = "test_resource";
-
-  /**
-   * AWS CloudFormation resource properties.
-   *
-   * This object is returned via cfnProperties
-   * @internal
-   */
-  protected readonly _properties: any;
-  constructor(scope: Construct, id: string, config: TestResourceConfig) {
-    super(scope, id, {
-      terraformResourceType: "test_resource",
-      terraformGeneratorMetadata: {
-        providerName: TestProviderMetadata.TYPE,
-        providerVersionConstraint: "~> 2.0",
-      },
-      provider: config.provider,
-      dependsOn: config.dependsOn,
-      count: config.count,
-      lifecycle: config.lifecycle,
-      provisioners: config.provisioners,
-      forEach: config.forEach,
-    });
-    this._properties = config.properties || {};
-  }
-
-  public get names(): string[] {
-    return this.getListAttribute("names");
-  }
-
-  protected synthesizeAttributes(): { [name: string]: any } {
-    return this._properties;
-  }
-}
