+++ test/aws/notify/rule.test.ts
@@ -5,13 +5,7 @@ import {
   cloudwatchEventTarget,
   cloudwatchEventBusPolicy,
 } from "@cdktf/provider-aws";
-import {
-  App,
-  Testing,
-  TerraformResource,
-  TerraformMetaArguments,
-  Lazy,
-} from "cdktf";
+import { App, Testing, Lazy } from "cdktf";
 import { Construct, IConstruct } from "constructs";
 import "cdktf/lib/testing/adapters/jest";
 import { AwsStack } from "../../../src/aws/aws-stack";
@@ -30,6 +24,7 @@ import { Rule } from "../../../src/aws/notify/rule";
 import { Duration } from "../../../src/duration";
 import { Fn } from "../../../src/terra-func";
 import { Annotations, Template } from "../../assertions";
+import { TestResource } from "../../test-resource";
 
 const ruleTfResource = cloudwatchEventRule.CloudwatchEventRule.tfResourceType;
 const ruleTargetTfResource =
@@ -1339,47 +1334,5 @@ class SomeTarget implements IRuleTarget {
       kinesisParameters: { partitionKeyPath: "partitionKeyPath" },
       targetResource: this.resource,
     };
-  }
-}
-
-export enum TestProviderMetadata {
-  TYPE = "test",
-}
-export interface TestResourceConfig extends TerraformMetaArguments {
-  readonly properties?: { [name: string]: any };
-}
-export class TestResource extends TerraformResource {
-  public static readonly tfResourceType: string = "test_resource";
-
-  /**
-   * AWS CloudFormation resource properties.
-   *
-   * This object is returned via cfnProperties
-   * @internal
-   */
-  protected readonly _properties: any;
-  constructor(scope: Construct, id: string, config: TestResourceConfig) {
-    super(scope, id, {
-      terraformResourceType: "test_resource",
-      terraformGeneratorMetadata: {
-        providerName: TestProviderMetadata.TYPE,
-        providerVersionConstraint: "~> 2.0",
-      },
-      provider: config.provider,
-      dependsOn: config.dependsOn,
-      count: config.count,
-      lifecycle: config.lifecycle,
-      provisioners: config.provisioners,
-      forEach: config.forEach,
-    });
-    this._properties = config.properties || {};
-  }
-
-  public get names(): string[] {
-    return this.getListAttribute("names");
-  }
-
-  protected synthesizeAttributes(): { [name: string]: any } {
-    return this._properties;
   }
 }
